Add AioMySql Support (may work, docs and other stuff not done)
