+++ pydapper/__init__.py
@@ -7,7 +7,8 @@ from .main import using
 from .main import using_async
 from .mssql import PymssqlCommands as _PymssqlCommands
 from .mysql import MySqlConnectorPythonCommands as _MySqlConnectorPythonCommands
-from .mariadb import MariaDbConnectorPythonCommands as _MariaDbConnectorPythonCommands
+from .mysql import AiomysqlCommands as _AiomysqlCommands
+from .mariadb import MariaDbCommands as _MariaDbCommands
 from .oracle import CxOracleCommands as _CxOracleCommands
 from .postgresql import AiopgCommands as _AioPgCommand
 from .postgresql import Psycopg2Commands as _Psycopg2Commands
