+++ pydapper/mysql/__init__.py
@@ -1,1 +1,2 @@
 from .mysql_connector_python import MySqlConnectorPythonCommands
+from .aiomysql import AiomysqlCommands
