+++ pyproject.toml
@@ -42,6 +42,7 @@ mariadb = { version = "^1.1.10", optional = true}
 cx-Oracle = { version = "^8.3.0", optional = true }
 oracledb = { version = "^1.1.1", optional = true }
 aiopg = { version = "^1.3.3", optional = true }
+aiomysql = { version = "^0.2.0", optional = true }
 coro-context-manager = "0.2.0"
 google-cloud-bigquery = {version = "^3.11.4", optional = true}
 google-cloud-bigquery-storage = {version = "^2.22.0", optional = true}
@@ -74,6 +75,8 @@ mysql-connector-python = ["mysql-connector-python"]
 cx_Oracle = ["cx-Oracle"]
 oracledb = ["oracledb"]
 aiopg = ["aiopg"]
+mariadb = ["mariadb"]
+aiomysql = ["aiomysql"]
 google-cloud-bigquery = ["google-cloud-bigquery"]
 # optional dep to make bigquery queries run faster
 google-cloud-bigquery-storage = ["google-cloud-bigquery-storage", "pyarrow", "numpy"]
@@ -86,6 +89,7 @@ all = [
     "cx_Oracle",
     "oracledb",
     "aiopg",
+    "aiomysql",
     "types-psycopg2",
     "types-pymssql",
     "google-cloud-bigquery",
